Terminer la 2eme partie de projet

+++ app/Http/Controllers/QuoteController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use App\Models\quote;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Auth;
 
 class QuoteController extends Controller
 {
@@ -12,7 +14,7 @@ class QuoteController extends Controller
      */
     public function index()
     {
-        $quotes = quote::all();
+        $quotes = Quote::with(['tags', 'categories'])->get();
         return response()->json($quotes);
     }
 
@@ -29,19 +31,30 @@ class QuoteController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
-            'author' => 'required',
-            'quote' => 'required',
-        ]);
+            $validatedData = $request->validate([
+                'author' => 'required',
+                'quote' => 'required',
+                'tags' => 'nullable|array|exists:tags,id',
+                'categories' => 'required|array|exists:categories,id',
+            ]);
+    
+            $validatedData['user_id'] = Auth::id();
+            $quote = quote::create($validatedData);
 
-        $quote = quote::create($validatedData);
-        return response()->json(
-            [
-                'message' => 'Quote created successfully',
-                'quote' => $quote
-            ],
-            201
-        );
+            if(request()->has('tags')){
+                $quote->tags()->sync($request->tags);
+            }
+            if(request()->has('categories')){
+                $quote->categories()->sync($request->categories);
+            }
+
+            return response()->json(
+                [
+                    'message' => 'Quote created successfully',
+                    'quote' => $quote
+                ],
+                201
+            );
     }
 
     /**
@@ -68,22 +81,26 @@ class QuoteController extends Controller
      */
     public function update(Request $request, quote $quote)
     {
-        $validatedData = $request->validate([
-            'author' => 'required',
-            'quote' => 'required',
-        ]);
-
-        $quote->author = $validatedData['author'];
-        $quote->quote = $validatedData['quote'];
-        $quote->save();
-
-        return response()->json(
-            [
-                'message' => 'Quote updated successfully',
-                'quote' => $quote
-            ],
-            200
-        );
+        if(Gate::Allows('CheckPermission', $quote)){
+            $validatedData = $request->validate([
+                'author' => 'required',
+                'quote' => 'required',
+            ]);
+    
+            $quote->author = $validatedData['author'];
+            $quote->quote = $validatedData['quote'];
+            $quote->save();
+    
+            return response()->json(
+                [
+                    'message' => 'Quote updated successfully',
+                    'quote' => $quote
+                ],
+                200
+            );
+        }else{
+            return response()->json(['message' => 'You are not allowed to update a quote'], 403);
+        }
     }
 
     /**
@@ -91,13 +108,17 @@ class QuoteController extends Controller
      */
     public function destroy(quote $quote)
     {
-        $quote->delete();
-        return response()->json(
-            [
-                'message' => 'Quote deleted successfully'
-            ],
-            200
-        );
+        if(Gate::Allows('CheckPermission', $quote)){
+            $quote->delete();
+            return response()->json(
+                [
+                    'message' => 'Quote deleted successfully'
+                ],
+                200
+            );
+        }else{
+            return response()->json(['message' => 'You are not allowed to delete a quote'], 403);
+        }
     }
 
 
@@ -120,6 +141,20 @@ class QuoteController extends Controller
 
     public function getRandom($nb){
         $quote = quote::inRandomOrder()->limit($nb)->get();
+        return response()->json($quote);
+    }
+
+    public function LikeQuote(Request $request){
+        $quote = quote::find($request->id);
+        $quote->like++;
+        $quote->save();
+        return response()->json($quote);
+    }
+
+    public function DislikeQuote(Request $request){
+        $quote = quote::find($request->id);
+        $quote->like--;
+        $quote->save();
         return response()->json($quote);
     }
